+++ Engine/Source/Developer/Cooker/Private/Commandlets/CookCommandlet.cpp
@@ -1,6 +1,5 @@
 #include "Commandlets/CookCommandlet.h"
 
-#include "Animation/CookedSkeletal.h"
 #include "CookRecipe.h"
 #include "StaticMeshBuilder.h"
 
@@ -12,23 +11,15 @@ namespace
 
 	void PrintUsage()
 	{
-		std::cout << "LeonCook - cook meshes / skeletal assets for Leon (UCookCommandlet)\n\n"
+		std::cout << "LeonCook - cook static meshes for Leon (UCookCommandlet)\n\n"
 				  << "Usage:\n"
 				  << "  LeonCook staticmesh --obj <mesh.obj>|--fbx <m.fbx>|--gltf <m.gltf> --out <m.lmesh>\n"
 				  << "    [--materials <dir>]  (glTF: write .lmat + textures)\n\n"
-				  << "  LeonCook character --name <Name> --mesh <idle.fbx> --run <run.fbx> --out <dir>\n"
-				  << "    [--jump <JumpingUp.fbx>] [--fall <FallingIdle.fbx>] [--land <Land.fbx>]\n\n"
-				  << "  LeonCook anim --fbx <clip.fbx> --skeleton <Bot.lskel>\n"
-				  << "    --name <ClipName> --out <Anims/Clip.lanim> [--noloop]\n\n"
 				  << "  LeonCook recipe <file.json>\n"
 				  << "    Runs steps from a recipe; relative paths resolve next to the JSON file.\n"
-				  << "    Step types: character | anim | staticmesh\n\n"
+				  << "    Step types: staticmesh\n\n"
 				  << "Writes (staticmesh):\n"
-				  << "  binary .lmesh (LMSH)\n"
-				  << "Writes (character):\n"
-				  << "  <Name>.lskel / .lskm / Materials / Anims / blendspace / .lchar\n"
-				  << "Writes (anim):\n"
-				  << "  <out>.lanim\n";
+				  << "  binary .lmesh (LMSH)\n";
 	}
 
 	[[nodiscard]] const char* ArgValue(int Argc, char** Argv, int& I)
@@ -152,153 +143,6 @@ namespace
 		return 0;
 	}
 
-	[[nodiscard]] int CookCharacter(int Argc, char** Argv)
-	{
-		std::string Name;
-		std::string MeshFbx;
-		std::string RunFbx;
-		std::string OutDir;
-		FCookJumpAnimPaths JumpAnims{};
-
-		for (int I = 2; I < Argc; ++I)
-		{
-			const std::string A = Argv[I];
-			if (A == "--name")
-			{
-				if (const char* V = ArgValue(Argc, Argv, I))
-				{
-					Name = V;
-				}
-			}
-			else if (A == "--mesh")
-			{
-				if (const char* V = ArgValue(Argc, Argv, I))
-				{
-					MeshFbx = V;
-				}
-			}
-			else if (A == "--run")
-			{
-				if (const char* V = ArgValue(Argc, Argv, I))
-				{
-					RunFbx = V;
-				}
-			}
-			else if (A == "--jump")
-			{
-				if (const char* V = ArgValue(Argc, Argv, I))
-				{
-					JumpAnims.JumpStartFbx = V;
-				}
-			}
-			else if (A == "--fall")
-			{
-				if (const char* V = ArgValue(Argc, Argv, I))
-				{
-					JumpAnims.FallLoopFbx = V;
-				}
-			}
-			else if (A == "--land")
-			{
-				if (const char* V = ArgValue(Argc, Argv, I))
-				{
-					JumpAnims.LandFbx = V;
-				}
-			}
-			else if (A == "--out")
-			{
-				if (const char* V = ArgValue(Argc, Argv, I))
-				{
-					OutDir = V;
-				}
-			}
-			else
-			{
-				std::cerr << "Unknown arg '" << A << "'\n";
-				return 1;
-			}
-		}
-
-		if (Name.empty() || MeshFbx.empty() || RunFbx.empty() || OutDir.empty())
-		{
-			std::cerr << "character requires --name --mesh --run --out\n";
-			PrintUsage();
-			return 1;
-		}
-
-		if (!CookCharacterFromFbx(Name, MeshFbx, RunFbx, OutDir, JumpAnims))
-		{
-			std::cerr << "Cook failed\n";
-			return 2;
-		}
-		return 0;
-	}
-
-	[[nodiscard]] int CookAnim(int Argc, char** Argv)
-	{
-		std::string Fbx;
-		std::string Skeleton;
-		std::string Name;
-		std::string OutJson;
-		bool bLooping = true;
-
-		for (int I = 2; I < Argc; ++I)
-		{
-			const std::string A = Argv[I];
-			if (A == "--fbx")
-			{
-				if (const char* V = ArgValue(Argc, Argv, I))
-				{
-					Fbx = V;
-				}
-			}
-			else if (A == "--skeleton")
-			{
-				if (const char* V = ArgValue(Argc, Argv, I))
-				{
-					Skeleton = V;
-				}
-			}
-			else if (A == "--name")
-			{
-				if (const char* V = ArgValue(Argc, Argv, I))
-				{
-					Name = V;
-				}
-			}
-			else if (A == "--out")
-			{
-				if (const char* V = ArgValue(Argc, Argv, I))
-				{
-					OutJson = V;
-				}
-			}
-			else if (A == "--noloop")
-			{
-				bLooping = false;
-			}
-			else
-			{
-				std::cerr << "Unknown arg '" << A << "'\n";
-				return 1;
-			}
-		}
-
-		if (Fbx.empty() || Skeleton.empty() || OutJson.empty())
-		{
-			std::cerr << "anim requires --fbx --skeleton --out\n";
-			PrintUsage();
-			return 1;
-		}
-
-		if (!CookAnimSequenceFromFbx(Fbx, Skeleton, OutJson, Name, bLooping))
-		{
-			std::cerr << "Cook anim failed\n";
-			return 2;
-		}
-		return 0;
-	}
-
 } // namespace
 
 int32 UCookCommandlet::Main(int32 Argc, char** Argv)
@@ -319,14 +163,6 @@ int32 UCookCommandlet::Main(int32 Argc, char** Argv)
 	if (Mode == "staticmesh")
 	{
 		return CookStaticMesh(Argc, Argv);
-	}
-	if (Mode == "character")
-	{
-		return CookCharacter(Argc, Argv);
-	}
-	if (Mode == "anim")
-	{
-		return CookAnim(Argc, Argv);
 	}
 	if (Mode == "recipe")
 	{